+++ src/targeted/targets/conditional/condition-target.mjs
@@ -1,7 +1,10 @@
 import { MODULE_NAME } from '../../../consts.mjs';
-import { keyValueSelect } from '../../../handlebars-handlers/bonus-inputs/key-value-select.mjs';
+import { radioInput } from '../../../handlebars-handlers/bonus-inputs/radio-input.mjs';
+import { traitInput } from '../../../handlebars-handlers/trait-input.mjs';
 import { currentTargetedActors } from '../../../util/get-current-targets.mjs';
+import { listFormat } from '../../../util/list-format.mjs';
 import { localize, localizeFluentDescription } from '../../../util/localize.mjs';
+import { truthiness } from '../../../util/truthiness.mjs';
 import { BaseConditionalTarget } from './_base-conditional.target.mjs';
 
 const targetChoices =  /** @type {const} */ ({
@@ -10,8 +13,17 @@ const targetChoices =  /** @type {const} */ ({
 });
 
 /**
- * @typedef {keyof typeof targetChoices} TargetOptions
- */
+ * @typedef {'self' | 'target'} TargetOption
+*/
+
+/** @type {Record<TargetOption, string>} */
+const labeledChoices = {
+    self: '',
+    target: '',
+};
+
+const all = 'all';
+const any = 'any';
 
 /**
  * @extends BaseConditionalTarget
@@ -23,6 +35,7 @@ export class ConditionTarget extends BaseConditionalTarget {
      */
     static get sourceKey() { return 'condition'; }
     static get #targetKey() { return `${this.key}-target`; }
+    static get #anyOrAllKey() { return `${this.key}-any-or-all`; }
 
     /**
      * @override
@@ -38,8 +51,9 @@ export class ConditionTarget extends BaseConditionalTarget {
      */
     static fluentDescription(source) {
         const key = this.#getTargetType(source) === 'self' ? 'condition-self' : 'condition-target';
-        const condition = source.getFlag(MODULE_NAME, this.key);
-        return localizeFluentDescription(key, { condition: condition ? pf1.registry.conditions.get(condition)?.name : '' });
+        const conditions = this.#getConditionsAslListText(source);
+        if (!conditions) return '';
+        return localizeFluentDescription(key, { conditions });
     }
 
     /**
@@ -48,8 +62,7 @@ export class ConditionTarget extends BaseConditionalTarget {
      */
     static init() {
         Hooks.once('ready', () =>
-            // @ts-ignore
-            Object.entries(targetChoices).forEach(([key, value]) => targetChoices[key] = localize(value))
+            (/** @type {[TargetOption, string][]} */ (Object.entries(targetChoices))).forEach(([key, value]) => labeledChoices[key] = localize(value))
         );
     }
 
@@ -59,12 +72,12 @@ export class ConditionTarget extends BaseConditionalTarget {
      * @returns {Nullable<string[]>}
      */
     static getHints(source) {
-        const condition = source.getFlag(MODULE_NAME, this.key);
+        const conditions = this.#getConditionsFromItem(source);
         const targetOrSelf = this.#getTargetType(source);
-        if (condition) {
-            const name = pf1.registry.conditions.get(condition)?.name;
-            if (name) {
-                return [`${name} - ${targetChoices[targetOrSelf]}`];
+        if (conditions.length) {
+            const conditionsText = this.#getConditionsAslListText(source);
+            if (conditionsText) {
+                return [`${labeledChoices[targetOrSelf]}: ${conditionsText}`];
             }
         }
     }
@@ -80,16 +93,17 @@ export class ConditionTarget extends BaseConditionalTarget {
         const currentTargets = currentTargetedActors();
 
         const bonusSources = sources.filter((source) => {
-            const condition = source.getFlag(MODULE_NAME, this.key);
-            if (!condition) return false;
+            const conditions = this.#getConditionsFromItem(source);
+            if (!conditions.length) return false;
 
             const targetOrSelf = this.#getTargetType(source);
+            const everyOrSome = source.getFlag(MODULE_NAME, this.#anyOrAllKey) === all ? 'every' : 'some';
             if (targetOrSelf === 'self') {
-                return actor.statuses.has(condition);
+                return conditions[everyOrSome]((condition) => actor.statuses.has(condition));
             }
             else {
                 if (!currentTargets.length) return false;
-                return currentTargets.every((a) => a.statuses.has(condition));
+                return currentTargets.every((a) => conditions[everyOrSome]((condition) => a.statuses.has(condition)));
             }
         });
 
@@ -101,7 +115,7 @@ export class ConditionTarget extends BaseConditionalTarget {
      * @override
      * @param {ItemPF} item
      * @param {keyof Conditions} condition
-     * @param {TargetOptions} [targetOrSelf]
+     * @param {TargetOption} [targetOrSelf]
      * @returns {Promise<void>}
      */
     static async configure(item, condition, targetOrSelf = 'target') {
@@ -116,6 +130,16 @@ export class ConditionTarget extends BaseConditionalTarget {
         });
     }
 
+    /** @type {Record<string, string>?} */
+    static #labeledConditions = null;
+    /** @returns {Record<string, string>} */
+    static get #getLabeledConditions() {
+        return this.#labeledConditions ||= pf1.registry.conditions.contents
+            .map(({ id, name }) => ({ key: id, label: name }))
+            .sort((a, b) => a.label.localeCompare(b.label))
+            .reduce((acc, { key, label }) => ({ ...acc, [key]: label }), {});
+    }
+
     /**
      * @inheritdoc
      * @override
@@ -126,10 +150,8 @@ export class ConditionTarget extends BaseConditionalTarget {
      * @param {ItemPF} options.item
      */
     static showInputOnItemSheet({ html, isEditable, item }) {
-        const choices = pf1.registry.conditions.contents
-            .map(({ id, name }) => ({ key: id, label: name }));
-        keyValueSelect({
-            choices,
+        traitInput({
+            choices: this.#getLabeledConditions,
             item,
             journal: this.journal,
             key: this.key,
@@ -140,25 +162,68 @@ export class ConditionTarget extends BaseConditionalTarget {
             inputType: 'target',
         });
 
-        keyValueSelect({
-            choices: targetChoices,
+        radioInput({
             item,
             journal: this.journal,
             key: this.#targetKey,
             parent: html,
             tooltip: this.tooltip,
+            values: Object.entries(labeledChoices).map(([key, label]) => ({ id: key, label })),
         }, {
             canEdit: isEditable,
             inputType: 'target',
             isSubLabel: true,
         });
+
+        const current = this.#getConditionsFromItem(item);
+        if (current.length > 1) {
+            const radioValues = [
+                { id: all, label: localize(`target-toggle.${all}`) },
+                { id: any, label: localize(`target-toggle.${any}`) },
+            ];
+
+            radioInput({
+                item,
+                journal: this.journal,
+                key: this.#anyOrAllKey,
+                values: radioValues,
+                parent: html,
+            }, {
+                canEdit: isEditable,
+                inputType: 'target',
+                isSubLabel: true,
+            });
+        }
     }
 
     /**
      * @param {ItemPF} source
-     * @returns {TargetOptions}
+     * @returns {TargetOption}
      */
     static #getTargetType(source) {
         return source.getFlag(MODULE_NAME, this.#targetKey) || 'target';
+    }
+
+    /**
+     * @param {ItemPF} source
+     * @returns {(keyof Conditions)[]}
+     */
+    static #getConditionsFromItem(source) {
+        return source.getFlag(MODULE_NAME, this.key) || [];
+    }
+
+    /**
+     * @param {ItemPF} source
+     * @returns {string}
+     */
+    static #getConditionsAslListText(source) {
+        const conditions = this.#getConditionsFromItem(source)
+            .map((c) => pf1.registry.conditions.get(c)?.name ?? '')
+            .filter(truthiness);
+        if (!conditions.length) return '';
+
+        const join = source.getFlag(MODULE_NAME, this.#anyOrAllKey) === all ? 'and' : 'or';
+        const text = listFormat(conditions, join);
+        return text;
     }
 }